+++ CHANGES.rst
@@ -105,6 +105,9 @@ New Features
   - ``extract_array`` now offers different options to deal with array
     boundaries [#3727]
 
+  - Added a new ``Cutout`` class to create postage stamp image cutouts
+    with optional WCS propagation. [#3823]
+
 - ``astropy.stats``
 
   - Added ``sigma_lower`` and ``sigma_upper`` keywords to